requirements.txt
alembic==1.14.0
aniso8601==9.0.1
attrs==24.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
Flask==2.2.5
flask-cors==5.0.1
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.3
Flask-Migrate==4.0.4
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
MarkupSafe==3.0.2
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.2.7
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
protobuf==3.20.3
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
